PY-18_BlogAPI/requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
Faker==18.9.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
